ClangImporter: inject modulemaps via the VFS
This is the second step towards removal of the content injection into
the Visual Studio installation.  With this the new path is enabled to
create the mappings to the SDK shared content.  Subsequent improvements
involve the introduction of new options to the Swift driver and
Swift frontend:
  1. `-windows-sdk-root`
  2. `-windows-sdk-version`
  3. `-visualc-tools-root`
  4. `-visualc-tools-version`
The introduction of these options will then feed into the path selection
here enabling granular control over the VCRuntime and Windows SDK by the
user.

Thanks to @artemcm for the very helpful discussion and help on
identifying a good set of options for the driver.  Thanks to @stevapple
for the inspiration for this patch.