Driver: encode the RSP mode for clang in the invocation
When building the `InvocationInfo` for the link job, we need to ensure
that we pass along the quoting mode for clang to properly decode the
response file. This is a problem when building with the early
swift-driver and cross-compiling where the emitted response file uses
the build encoding but conditions the behaviour on the host encoding.